requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Pillow>=12.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
reportlab==4.0.7
